feat(refactor): move and rename fragment classes
